A few initial changes to start working on protoss support for the macro manager, however I have yet to implement the main changes that will be needed:
I think the main change will need to be a way to have a task run and complete with no worker specified (since a building task should be able to release a worker once the building starts warping in). Also I'll need to add a way to fork off tasks when multiple building tasks are in a task stream so that a probe can immediately start the next building once the first building in the task stream is done. Training units and researching upgrades and abilities probably won't need any changes.